Refactor boot modules and enhance RISC-V initialization
- Removed unused early_mmu and entry module imports in AArch64 boot code.
- Implemented init_boot_cpu and trap_init functions in RISC-V boot module for CPU initialization and trap handling.
- Moved trap_init function from riscv64/mod.rs to riscv64/boot/mod.rs for better organization.
- Updated Limine boot requests to public visibility for broader access.
- Created a new lib.rs file to define the Scarlet kernel structure and initialization process.
- Enhanced the start_kernel function to include comprehensive boot information handling and initialization steps.
- Improved memory management and device discovery during the boot process.
- Added detailed logging for kernel initialization stages and device management.
