+++ bsp/aarch64-limine/targets/aarch64-unknown-none-elf.json
@@ -1,0 +1,32 @@
+{
+  "arch": "aarch64",
+  "code-model": "large",
+  "cpu": "generic",
+  "crt-objects-fallback": "false",
+  "data-layout": "e-m:e-p270:32:32-p271:32:32-p272:64:64-i8:8:32-i16:16:32-i64:64-i128:128-n32:64-S128-Fn32",
+  "eh-frame-header": false,
+  "emit-debug-gdb-scripts": false,
+  "features": "+strict-align",
+
+  "linker": "rust-lld",
+  "linker-flavor": "gnu-lld",
+  "llvm-target": "aarch64-elf",
+  "llvm-abiname": "lp64",
+  "max-atomic-width": 64,
+  "metadata": {
+    "description": "Bare AArch64 (ARMv8-A ISA)",
+    "host_tools": false,
+    "std": false,
+    "tier": 2
+  },
+  "panic-strategy": "abort",
+  "pre-link-args": {
+      "ld.lld": ["-T", "lds/aarch64_limine.ld"]
+    },
+  "relocation-model": "static",
+  "supported-sanitizers": [
+    "shadow-call-stack",
+    "kernel-address"
+  ],
+  "target-pointer-width": 64
+}
