+++ bsp/riscv64-limine/.cargo/config.toml
@@ -1,0 +1,16 @@
+[profile.dev]
+opt-level = 3
+
+[profile.test]
+opt-level = 3
+
+[target.riscv64gc-unknown-none-elf]
+runner = "tools/run.sh"
+
+[build]
+target = "targets/riscv64gc-unknown-none-elf.json"
+
+[unstable]
+build-std = ["core", "compiler_builtins", "alloc"]
+build-std-features = ["compiler-builtins-mem"]
+unstable-options = true
