+++ bsp/riscv64-limine/.gitignore
@@ -1,0 +1,1 @@
+target
