+++ kernel/src/arch/aarch64/boot/mod.rs
@@ -4,13 +4,6 @@
 //! This module provides the entry points and initialization code for
 //! starting the kernel on AArch64 systems.
 
-mod early_mmu;
-mod entry;
-
-// Re-export the entry points for the linker
-pub use early_mmu::*;
-pub use entry::*;
-
 /// Initialize AArch64-specific boot components
 ///
 /// This function is called during kernel initialization to set up
